source/requirements.txt
# Framework
Django==1.10

# Libs 
djangorestframework==3.4.6
djangorestframework-jwt==1.8
django-rest-auth==0.8.1
django-allauth==0.27.0
django-filter==0.14.0
mysqlclient==1.3.7
django-compressor==2.1

# Server
gunicorn==19.6.0

#redis==2.10.3


